[댓글 기능] 유니 미션 제출합니다 (#126)
* feat : article user 관계 매핑

* html : article 작성자 표시

* feat : article 수정 본인만

* feat : article 삭제 본인만

* feat : comment 추가

* refactor : comment 추가

* feat : comment 수정, 삭제

* feat : comment 수정, 삭제 테스트

* refactor : comment service 분리

* refactor : test 오류 수정

* refactor : createDate 적용

* refactor : service 분리

* refactor : 다른 도메인이면 서비스를 이용하도록 수정

* refactor : field injection -> constructor injection 수정

* refactor : userDto 분리

* refactor : userDto 분리

* refactor : import 정리

* refactor : test 수정

* refactor : test 수정

* refactor : column 추가

* refactor : column 추가

* refactor : test 수정

+++ src/main/java/techcourse/myblog/service/ArticleService.java
@@ -6,18 +6,27 @@ import techcourse.myblog.domain.article.Article;
 import techcourse.myblog.domain.article.ArticleDto;
 import techcourse.myblog.domain.article.ArticleDtos;
 import techcourse.myblog.domain.article.ArticleRepository;
+import techcourse.myblog.domain.user.UserDto;
 
 import javax.transaction.Transactional;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 @Service
 public class ArticleService {
-    @Autowired
-    private ArticleRepository articleRepository;
+    private final ArticleRepository articleRepository;
 
-    public long createArticle(ArticleDto articleDto) {
+    private final UserService userService;
+
+    @Autowired
+    public ArticleService(ArticleRepository articleRepository, UserService userService) {
+        this.articleRepository = articleRepository;
+        this.userService = userService;
+    }
+
+    public long createArticle(ArticleDto articleDto, UserDto userDto) {
+        UserDto findUserDto = userService.findByUserEmail(userDto);
+        articleDto.setUserDto(findUserDto);
         Article article = articleRepository.save(articleDto.toEntity());
         return article.getId();
     }
@@ -29,8 +38,9 @@ public class ArticleService {
     }
 
     @Transactional
-    public ArticleDto updateByArticle(long articleId, ArticleDto articleDto) {
+    public ArticleDto updateByArticle(long articleId, ArticleDto articleDto, UserDto userDto) {
         Optional<Article> maybeArticle = articleRepository.findById(articleId);
+        checkAuthor(articleId, userDto);
         if (maybeArticle.isPresent()) {
             maybeArticle.get().update(articleDto.toEntity());
 
@@ -39,7 +49,9 @@ public class ArticleService {
         throw new IllegalArgumentException("업데이트 할 수 없습니다.");
     }
 
-    public void deleteById(long articleId) {
+    @Transactional
+    public void deleteById(long articleId, UserDto userDto) {
+        checkAuthor(articleId, userDto);
         articleRepository.deleteById(articleId);
     }
 
@@ -49,5 +61,14 @@ public class ArticleService {
 
     public List<ArticleDto> readByCategoryId(long categoryId) {
         return new ArticleDtos(articleRepository.findByCategoryId(categoryId)).getArticleDtos();
+    }
+
+    private void checkAuthor(long articleId, UserDto userDto) {
+        UserDto findUserDto = userService.findByUserEmail(userDto);
+        articleRepository.findById(articleId).ifPresent(article -> {
+            if (!findUserDto.toEntity().checkAuthor(article.getAuthor().getId())) {
+                throw new IllegalArgumentException("허가되지 않은 사용자입니다.");
+            }
+        });
     }
 }
